Post Sprint 1 Commit.
There are a number of things I need to implement, and in truth I didn't really finish sprint 1. I didn't actually get the model relationships correct, redirects are a mess, there is zero visual styling implemented, both players and rounds can be seen in a list (neither of which should be), landing page is the player list page (when it should be a sign-in or sign-up page).